Add files via upload
This is the first object-oriented Python code. It shows creation of two rows of nodes for a 1-D CVM. There are four nodes in each row, replicating Fig. 3 of "The Cluster Variation Method: A Primer for Neuroscientists," available at: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5187558/. 
This code shows only the activation values for each node, and the "w" values (the activations of the nodes to the immediate left and right of a given node).